% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accumulation.R
\name{autoplot.Accumulation}
\alias{autoplot.Accumulation}
\title{Plot Diversity Accumulation}
\usage{
\method{autoplot}{Accumulation}(
  object,
  ...,
  q = 0,
  main = "Accumulation of ...",
  xlab = "Sample size...",
  ylab = "Diversity...",
  ylim = NULL,
  lineH0 = TRUE,
  ShadeColor = "grey75",
  BorderColor = "red"
)
}
\arguments{
\item{object}{An "Accumulation" object that cat be accumulation of diversity (\link{DivAccum}), entropy (\link{EntAccum}) or the Mixing index (\link{Mixing}).}

\item{...}{Further plotting arguments.}

\item{q}{The order of Diversity}

\item{main}{Main title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{Limits of the Y-axis, as a vector of two numeric values.}

\item{lineH0}{if \code{TRUE}, the values of the null hypothesis are plotted.}

\item{ShadeColor}{The color of the shaded confidence envelope.}

\item{BorderColor}{The color of the borders of the confidence envelope.}
}
\description{
Plot Diversity Accumulation
}
\examples{
# Generate a random community
spCommunity <- rSpCommunity(1, size=50, S=3)
# Calculate mixing indices of order 0 and 1
accum <- Mixing(spCommunity, q.seq=c(0,1))
autoplot(accum, q=0)
}
