% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{CheckSpatDivArguments}
\alias{CheckSpatDivArguments}
\title{CheckSpatDivArguments}
\usage{
CheckSpatDivArguments()
}
\value{
Returns \code{TRUE} or stops if a problem is detected.
}
\description{
Checks the arguments of a function of the package SpatDiv
}
\details{
The function compares the arguments passed to its parent function to the type they should be and performs some extra tests (\emph{e.g.} probabilities must be positive and sum to 1). It stops if an argument is not correct.
}
\author{
Eric Marcon <Eric.Marcon@ecofog.gf>
}
\keyword{internal}
