% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Diversity.wmppp}
\alias{Diversity.wmppp}
\title{HCDT diversity of a spatialized community}
\usage{
\method{Diversity}{wmppp}(NorP, q = 1, Correction = "Best",
  CheckArguments = TRUE, Ps = NULL, Ns = NULL)
}
\arguments{
\item{Correction}{A string containing one of the possible corrections to estimate a probability distribution: \code{"None"} (no correction, the default value), \code{"Chao2013"}, \code{"Chao2015"} or \code{"ChaoShen"}.}

\item{CheckArguments}{Logical; if \code{TRUE}, the function arguments are verified. Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated diversity. The name is that of the bias correction used.
}
\description{
\code{Diversity} calls \code{\link{Tsallis}} to calculate entropy and transforms it into diversity by calculating its deformed exponential.
}
\examples{
TODO
}
