% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntAccum.R
\name{EntAccum}
\alias{EntAccum}
\title{Entropy Accumulation}
\usage{
EntAccum(spCommunity, q.seq = seq(0, 2, by = 0.1), Correction = "None",
  n.seq = 1:ceiling(spCommunity$n/10), r.seq = NULL,
  CheckArguments = TRUE)
}
\arguments{
\item{spCommunity}{A spatialized community (An object of class "wmppp" (\code{\link{wmppp.object}}), with \code{PointType} values as species names.)}

\item{q.seq}{A numeric vector: the sequence of diversity orders to address. Default is from 0 to 2.}

\item{Correction}{A string containing one of the possible corrections to estimate a probability distribution: \code{"None"} (no correction, the default value), \code{"Chao2013"}, \code{"Chao2015"} or \code{"ChaoShen"}.}

\item{CheckArguments}{Logical; if \code{TRUE}, the function arguments are verified. Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{n}{A vector of integers. Entropy will be accumulated along this number of neighbors around each individual. Default is 10\% of the individuals.}

\item{r}{A vector of distances. If \code{NULL} accumulation is along \code{n}, else neighbors are accumulated in circles of radius \code{r}.}
}
\value{
A matrix containing average entropy. Columns
}
\description{
Entropy Accumulation
}
\examples{
TODO
}
