% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{Paracou6}
\alias{Paracou6}
\title{Paracou plot 6}
\format{
An object of class \link[dbmss:wmppp]{dbmss::wmppp}.
}
\source{
Permanent data census of Paracou: \url{http://paracou.cirad.fr/}
}
\usage{
Paracou6
}
\description{
This dataset is from Paracou field station, French Guiana, managed by \href{http://www.cirad.fr}{Cirad}.
It contains the position, species and basal area of all trees above 10 cm diameter at breast height (DBH) in a 6.25ha plot.
}
\keyword{datasets}
