% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntAccum.R
\name{plot.Accumulation}
\alias{plot.Accumulation}
\title{Plot Diversity Accumulation}
\usage{
\method{plot}{Accumulation}(x, ..., q = 0, type = "l",
  main = "Accumulation of ...", xlab = "Sample size",
  ylab = "Diversity...", ylim = NULL, LineWidth = 2,
  ShadeColor = "grey75", BorderColor = "red")
}
\arguments{
\item{...}{Further plotting arguments.}

\item{q}{The order of Diversity}

\item{type}{Plotting parameter. Default is "l".}

\item{main}{Main title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{Limits of the Y-axis, as a vector of two numeric values.}

\item{LineWidth}{Width of the Diversity Accumulation Curve line.}

\item{ShadeColor}{The color of the shaded confidence envelope.}

\item{BorderColor}{The color of the borders of the confidence envelope.}

\item{An}{\code{\link{Accumulation}} object that cat be accumulation of diversity (\code{\link{DivAccum}}), entropy (\code{\link{EntAccum}}) or the Mixing index (\code{\link{Mixing}}).}

\item{LineWidth}{Width of the Diversity Accumulation Curve line.}
}
\description{
Plot Diversity Accumulation
}
\examples{
TODO
}
