% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntAccum.R
\name{plot.DivAccum}
\alias{plot.DivAccum}
\title{Plot Diversity Accumulation}
\usage{
\method{plot}{DivAccum}(x, ..., q = 0, type = "l",
  main = paste("Accumulation of Diversity of Order", q),
  xlab = "Sample size", ylab = "Diversity", ylim = NULL, LineWidth = 2,
  ShadeColor = "grey75", BorderColor = "red")
}
\arguments{
\item{...}{Further plotting arguments.}

\item{q}{The order of Diversity}

\item{type}{Plotting parameter. Default is "l".}

\item{main}{Main title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{Limits of the Y-axis, as a vector of two numeric values.}

\item{LineWidth}{Width of the Diversity Accumulation Curve line.}

\item{ShadeColor}{The color of the shaded confidence envelope.}

\item{BorderColor}{The color of the borders of the confidence envelope.}

\item{A}{\code{\link{DivAccum}} object}

\item{LineWidth}{Width of the Diversity Accumulation Curve line.}
}
\description{
Plot Diversity Accumulation
}
\examples{
TODO
}
