% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Richness.character}
\alias{Richness.character}
\title{Richness of a vector of individuals}
\usage{
\method{Richness}{character}(NorP, Correction = "Chao1", Alpha = 0.05,
  JackOver = FALSE, ..., CheckArguments = TRUE)
}
\arguments{
\item{NorP}{A vector of factors.}

\item{Correction}{A string containing one of the possible corrections: \code{"None"} (no correction), \code{"ChaoShen"}, \code{"GenCov"}, \code{"Grassberger"}, \code{"Holste"}, \code{"Bonachela"}, \code{"ZhangGrabchak"}, or \code{"ChaoWangJost"}, \code{"Marcon"}, \code{"UnveilC"}, \code{"UnveiliC"}, \code{"UnveilJ"} or \code{"Best"}, the default value.  Currently, \code{"Best"} is \code{"ChaoWangJost"}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated diversity. The name is that of the bias correction used.
}
\description{
\code{Richness} is the number of species.
}
\examples{
#TODO
}
