% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSpCommunity.R
\name{rSpCommunity}
\alias{rSpCommunity}
\title{Random Spatialized Community}
\usage{
rSpCommunity(n, size = sum(NorP), NorP = 1, BootstrapMethod = "Chao2015",
  S = 300, Distribution = "lnorm", sd = 1, prob = 0.1, alpha = 40,
  Spatial = "Binomial", win = spatstat::owin(), Species,
  Sizes = "Uniform", CheckArguments = TRUE)
}
\arguments{
\item{n}{The number of communities to draw.}

\item{size}{The number of individuals to draw in each community.}

\item{NorP}{A numeric vector or a two-column matrix. Contains either abundances or probabilities. Two-column matrices should contain the observed abundances (or probabilities) in the first column and the expected ones in the second column, to allow using beta diversity functions.}

\item{BootstrapMethod}{The method used to obtain the probabilities to generate bootstrapped communities from observed abundances. If \code{"Marcon"}, the probabilities are simply the abundances divided by the total number of individuals (Marcon \emph{et al.}, 2012). If \code{"Chao2013"} or \code{"Chao2015"} (by default), a more sophisticated approach is used (see \code{\link{as.ProbaVector}}) following Chao \emph{et al.} (2013) or Chao \emph{et al.} (2015).}

\item{S}{The number of species.}

\item{Distribution}{The distribution of species frequencies. May be \code{"lnorm"} (log-normal), \code{"lseries"} (log-series), \code{"geom"} (geometric) or \code{"bstick"} (broken stick).}

\item{sd}{The simulated distribution standard deviation. For the log-normal distribution, this is the standard deviation on the log scale.}

\item{prob}{The probability of success in each trial.}

\item{alpha}{Fisher's alpha.}

\item{Spatial}{TODO}

\item{win}{TODO}

\item{Species}{TODO}

\item{Sizes}{TODO}

\item{CheckArguments}{Logical; if \code{TRUE}, the function arguments are verified. Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}
}
\value{
An object of class "wmppp" (\code{\link{wmppp.object}}), with \code{PointType} values as species names if \code{n}=1.
        An object of class "SpCommunities", which is a list of \code{\link{wmppp.object}}s, is returned if \code{n}>1.
}
\description{
Random Spatialized Community
}
\examples{
TODO
}
