% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpCommunity.R
\name{rSpSpecies}
\alias{rSpSpecies}
\title{Random Spatialized Distribution of a Species}
\usage{
rSpSpecies(
  n,
  Spatial = "Binomial",
  scale = 0.2,
  mu = 10,
  win = spatstat.geom::owin(),
  Species = NULL,
  Sizes = "Uniform",
  MinSize = 1,
  MaxSize = 1,
  Wscale = 20,
  shape = 2,
  CheckArguments = TRUE
)
}
\arguments{
\item{n}{The number of individuals to draw.}

\item{Spatial}{The spatial distribution of points.
May be "Binomial" (a completely random point pattern except for its fixed number of points) or
"Thomas" for a clustered point pattern with parameters \code{scale} and \code{mu}.}

\item{scale}{In Thomas point patterns, the standard deviation of random displacement (along each coordinate axis) of a point from its cluster center.}

\item{mu}{In Thomas point patterns, the mean number of points per cluster.
The intensity of the Poisson process of cluster centers is calculated as the number of points (\code{size}) per area divided by \code{mu}.}

\item{win}{The window containing the point pattern. It is an \link{owin} object.}

\item{Species}{A vector of characters or of factors containing the possible species.}

\item{Sizes}{The distribution of point sizes.
May be "Uniform" for a uniform distribution between \code{MinSize} and \code{MaxSize}.
By default, all sizes are 1.
May be "Weibull" with parameters \code{MinSize}, \code{Wscale} and \code{shape}.}

\item{MinSize}{The minimum size in a uniform or Weibull distribution.}

\item{MaxSize}{The maximum size in a uniform distribution.}

\item{Wscale}{The scale parameter in a Weibull distribution.}

\item{shape}{The shape parameter in a Weibull distribution.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A \link{wmppp.object}.
}
\description{
Random Spatialized Distribution of a Species
}
\examples{
spSpecies <- rSpSpecies(50, Spatial = "Thomas")
autoplot(spSpecies)

}
