% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Shannon.character}
\alias{Shannon.character}
\title{Shannon Entropy of a vector of individuals}
\usage{
\method{Shannon}{character}(NorP, Correction = "Best", ...,
  CheckArguments = TRUE)
}
\arguments{
\item{NorP}{A vector of characters.}

\item{Correction}{A string containing one of the possible corrections: see \code{\link{Shannon}}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated entropy. The name is that of the bias correction used.
}
\description{
Calculates the Shannon entropy of a vector of individuals.
}
\examples{
#TODO
}
