% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Simpson.factor}
\alias{Simpson.factor}
\title{Simpson Entropy of a vector of individuals}
\usage{
\method{Simpson}{factor}(NorP, Correction = "Lande", ..., CheckArguments = TRUE)
}
\arguments{
\item{NorP}{A vector of factors.}

\item{Correction}{A string containing one of the possible corrections: see \link{Simpson}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated entropy. The name is that of the bias correction used.
}
\description{
Calculates the Simpson entropy of a vector of individuals.
}
\examples{
Simpson(Paracou6$marks$PointType)

}
