% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Simpson.wmppp}
\alias{Simpson.wmppp}
\title{Simpson Entropy of a spatialized community}
\usage{
\method{Simpson}{wmppp}(NorP, Correction = "Lande", ..., CheckArguments = TRUE)
}
\arguments{
\item{NorP}{An object of class \link[dbmss:wmppp]{dbmss::wmppp}, with \code{PointType} values as species names.}

\item{Correction}{A string containing one of the possible corrections: see \link{Simpson}.}

\item{...}{Further arguments. Unsused.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
A named number equal to the calculated entropy. The name is that of the bias correction used.
}
\description{
Calculates the Simpson entropy of a probability vector.
}
\examples{
Simpson(Paracou6)

}
