% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\docType{package}
\name{SpatDiv}
\alias{SpatDiv}
\alias{SpatDiv-package}
\title{SpatDiv}
\description{
Spatially Explicit Measures of Diversity
}
\details{
This package extends the \emph{entropart} package \insertCite{Marcon2014c}{SpatDiv}.
It provides spatially explicit measures of diversity such as the mixing index.
}
\references{
\insertAllCited{}
}
