% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesDistribution.R
\name{SpeciesDistributions}
\alias{SpeciesDistributions}
\alias{as.SpeciesDistribution.wmppp}
\alias{as.SpeciesDistribution.factor}
\alias{as.SpeciesDistribution.character}
\alias{as.ProbaVector.wmppp}
\alias{as.ProbaVector.factor}
\alias{as.ProbaVector.character}
\alias{as.AbdVector.wmppp}
\alias{as.AbdVector.factor}
\alias{as.AbdVector.character}
\title{Species Distribution}
\usage{
\method{as.SpeciesDistribution}{wmppp}(x, ...)

\method{as.SpeciesDistribution}{factor}(x, ...)

\method{as.SpeciesDistribution}{character}(x, ...)

\method{as.ProbaVector}{wmppp}(x, ...)

\method{as.ProbaVector}{factor}(x, ...)

\method{as.ProbaVector}{character}(x, ...)

\method{as.AbdVector}{wmppp}(x, ...)

\method{as.AbdVector}{factor}(x, ...)

\method{as.AbdVector}{character}(x, ...)
}
\arguments{
\item{x}{An object of class "wmppp" (\code{\link{wmppp.object}}), with \code{PointType} values as species names, or a vector of factors or characters containing species names of each individual.}

\item{...}{Further arguments. Unsused.}
}
\value{
A vector of species abundances (\code{\link{AbdVector}}) or probabilities (\code{\link{ProbaVector}}).
}
\description{
A Species Distribution is a (preferably named) vector containing species abundances or probabilities.
\code{SpeciesDistribution} objects include \code{AbdVector} and \code{ProbaVector} objects.
}
\details{
\code{as.AbdVector} counts the number of individuals (points) per species (in \code{marks$PointType}).

\code{as.ProbaVector} normalizes the vector so that it sums to 1. If \code{Correction} is not \code{"None"}, the observed abundance distribution is used to estimate the actual species distribution. The list of species will be changed: zero-abundance species will be cleared, and some unobserved species will be added. First, observed species probabilities are estimated folllowing \insertCite{Chao2003;textual}{SpatDiv}, \emph{i.e.} input probabilities are multiplied by the sample coverage, or according to more sophisticated models: \insertCite{Chao2013;textual}{SpatDiv}, single-parameter model or \insertCite{Chao2015;textual}{SpatDiv}, two-parameter model. The total probability of observed species equals the sample coverage. Then, the distribution of unobserved species can be unveiled: their number is estimated according to \code{RCorrection} (if the Jackknife estimator is chosen, the \code{JackOver} argument allows using the order immediately over the optimal one). The coverage deficit (1 minus the sample coverage) is shared by the unobserved species equally (\code{Unveiling = "unif"}, \insertCite{Chao2013;textual}{SpatDiv} or according to a geometric distribution (\code{Unveiling = "geom"}, \insertCite{Chao2015;textual}{SpatDiv}.

\code{SpeciesDistribution} objects can be plotted. The \code{plot} method returns the estimated parameters of the fitted distribution. The broken stick has no parameter, so the maximum abundance is returned.
}
\references{
\insertAllCited{}
}
