% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropart.R
\name{Tsallis.factor}
\alias{Tsallis.factor}
\title{Tsallis (HCDT) Entropy of a vector of individuals}
\usage{
\method{Tsallis}{factor}(NorP, q = 1, Correction = "Best", ...,
  CheckArguments = TRUE)
}
\arguments{
\item{NorP}{A vector of factors.}

\item{q}{A number: the order of entropy. Some corrections allow only a positive number. Default is 1 for Shannon entropy.}

\item{Correction}{A string containing one of the possible corrections: \code{"None"} (no correction), \code{"ChaoShen"}, \code{"GenCov"}, \code{"Grassberger"}, \code{"Holste"}, \code{"Bonachela"}, \code{"ZhangGrabchak"}, or \code{"ChaoWangJost"}, \code{"Marcon"}, \code{"UnveilC"}, \code{"UnveiliC"}, \code{"UnveilJ"} or \code{"Best"}, the default value.  Currently, \code{"Best"} is \code{"ChaoWangJost"}.}

\item{...}{Further arguments. Unsused.}
}
\value{
A named number equal to the calculated entropy. The name is that of the bias correction used.
}
\description{
Calculates the HCDT, also known as Tsallis entropy of order \eqn{q} of a probability vector.
}
\details{
Tsallis (Havrda and Charvat, 1967; Daroczy, 1970; Tsallis, 1988) generalized entropy is a generalized measure of diversity (Jost, 2006).
See \code{\link{Tsallis}} for more details.
}
\examples{
#TODO
}
