% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accumulation.R
\name{DivAccum}
\alias{DivAccum}
\title{Diversity Accumulation}
\usage{
DivAccum(
  spCommunity,
  q.seq = seq(0, 2, by = 0.1),
  divCorrection = "None",
  n.seq = 1:ceiling(spCommunity$n/2),
  r.seq = NULL,
  spCorrection = "None",
  H0 = "None",
  Alpha = 0.05,
  NumberOfSimulations = 100,
  Individual = FALSE,
  ShowProgressBar = TRUE,
  CheckArguments = TRUE
)
}
\arguments{
\item{spCommunity}{A spatialized community (A \link{wmppp.object} with \code{PointType} values as species names.)}

\item{q.seq}{A numeric vector: the sequence of diversity orders to address. Default is from 0 to 2.}

\item{divCorrection}{A string containing one of the possible corrections to calculate diversity, see \link{Tsallis} for all possible values. \code{"None"} uses the plugin estimator. \code{"Best"} is the default value.}

\item{n.seq}{A vector of integers. Entropy will be accumulated along this number of neighbors around each individual. Default is 10\% of the individuals.}

\item{r.seq}{A vector of distances. If \code{NULL} accumulation is along \code{n}, else neighbors are accumulated in circles of radius \code{r}.}

\item{spCorrection}{The edge-effect correction to apply when estimating the entropy of a neighborhood community that does not fit in the window.
Does not apply if neighborhoods are defined by the number of neighbors. Default is "None".
"Extrapolation" extrapolates the observed diversity up to the number of individuals estimated in the full area of the neighborhood.}

\item{H0}{The null hypothesis to compare the distribution of \code{spCommunity} to. If "none", the default value, no null hypothesis is tested.
"Multinomial" means the community will be rarefied down to the number of neighbors of \code{n.seq}.
"RandomLocation" means the points will we randomly permuted accross their actual locations.
"Binomial" means the points will we uniformly and independently drawn in the window (a binomial point process is a Poisson point process conditionally to the number of points).}

\item{Alpha}{The risk level of the envelope of the null hypothesis. Default is 5\%.}

\item{NumberOfSimulations}{The number of bootstraps to build confidence intervals. Default is 100.}

\item{Individual}{If \code{TRUE}, individual neighborhood entropies are returned.}

\item{ShowProgressBar}{If \code{TRUE} (default), a progress bar is shown.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified. Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}
}
\value{
An "Accumulation" object that is a 3-D array containing average diversity.
The third dimension of the array is only of length 4: it contains observed diversity, its value under the null hypothesis, and the lower of upper bounds of that value.
The first two dimensions are respectively for $q$ values and the number of points of the neighborhood, starting from 1 (the point itself, with no neighbor).
}
\description{
Diversity Accumulation
}
\examples{
# Generate a random community
spCommunity <- rSpCommunity(1, size=50, S=3)
# Calculate the species accumulation curve and accumulation of Simpson diversity
accum <- DivAccum(spCommunity, q.seq=c(0,2))
plot(accum, q=0)
plot(accum, q=2)

}
