% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlot.R
\name{MapPlot}
\alias{MapPlot}
\alias{MapPlot.Accumulation}
\title{MapPlot}
\usage{
MapPlot(x, ...)

\method{MapPlot}{Accumulation}(x, q, NeighborHood, Nbx = 128, Nby = 128,
  Contour = TRUE, Palette = topo.colors(128, alpha = 1),
  ContourColor = "dark red", ReturnObject = FALSE, ...,
  CheckArguments = TRUE)
}
\arguments{
\item{x}{An object to map.}

\item{...}{Further parameters.}
}
\value{
\code{MapPlot.Accumulation} returns an \code{\link{autoKrige}} object that can be used to produce alternative maps.
}
\description{
Methods to map spatialized communities.
}
