% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlot.R
\name{MapPlot}
\alias{MapPlot}
\alias{MapPlot.Accumulation}
\alias{MapPlot.fv}
\title{MapPlot}
\usage{
MapPlot(x, ...)

\method{MapPlot}{Accumulation}(
  x,
  Order,
  NeighborHood,
  AllowJitter = TRUE,
  Nbx = 128,
  Nby = 128,
  Contour = TRUE,
  Palette = grDevices::topo.colors(128, alpha = 1),
  Contournlevels = 10,
  Contourcol = "dark red",
  ...,
  CheckArguments = TRUE
)

\method{MapPlot}{fv}(
  x,
  spCommunity,
  r,
  AllowJitter = TRUE,
  Nbx = 128,
  Nby = 128,
  Contour = TRUE,
  Palette = grDevices::topo.colors(128, alpha = 1),
  Contournlevels = 10,
  Contourcol = "dark red",
  ...,
  CheckArguments = TRUE
)
}
\arguments{
\item{x}{An object to map.}

\item{...}{Further parameters.}

\item{Order}{The order of diversity, i.e. the value of $q$.}

\item{NeighborHood}{The neighborhood size, i.e. the number of neighbors or the distance to consider.}

\item{AllowJitter}{If \code{TRUE}, duplicated points are jittered to avoid their elimination by the kriging procedure.}

\item{Nbx}{The number of columns (pixels) of the resulting map, 128 by default.}

\item{Nby}{The number of rows (pixels) of the resulting map, 128 by default.}

\item{Contour}{If \code{TRUE}, contours are added to the map.}

\item{Palette}{The color palette of the map.}

\item{Contournlevels}{The number of levels of contours.}

\item{Contourcol}{The color of the contour lines.}

\item{CheckArguments}{If \code{TRUE} (default), the function arguments are verified.
Should be set to \code{FALSE} to save time in simulations for example, when the arguments have been checked elsewhere.}

\item{spCommunity}{A spatialized community (A \link{wmppp.object} with \code{PointType} values as species names.)}

\item{r}{The distance at which the function value must be considered.}
}
\value{
An \link{autoKrige} object that can be used to produce alternative maps.
}
\description{
Map Spatialized Communities
}
\details{
Map an "Accum" object.
}
\examples{
# Generate a random community
spCommunity <- rSpCommunity(1, size=50, S=10)
# Calculate the species accumulation curve
accum <- DivAccum(spCommunity, q.seq=c(0,1), n.seq=4, Individual=TRUE)
# Plot the local richness, accumulated up to 5 individuals.
MapPlot(accum, Order=0, NeighborHood=5)

# Generate a random community
spCommunity <- rSpCommunity(1, size=50, S=9, Species = paste0("sp",1:9))
library("dbmss")
M_i <- Mhat(spCommunity, ReferenceType = "sp1", Individual = TRUE)
MapPlot(M_i, spCommunity[spCommunity$marks$PointType=="sp1"], r=0.5)

}
