% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accumulation.R
\name{plot.Accumulation}
\alias{plot.Accumulation}
\title{Plot Diversity Accumulation}
\usage{
\method{plot}{Accumulation}(
  x,
  ...,
  q = dimnames(x$Accumulation)$q[1],
  type = "l",
  main = "Accumulation of ...",
  xlab = "Sample size...",
  ylab = "Diversity...",
  ylim = NULL,
  lineH0 = TRUE,
  LineWidth = 2,
  ShadeColor = "grey75",
  BorderColor = "red"
)
}
\arguments{
\item{x}{An "Accumulation" object that cat be accumulation of diversity (\code{\link{DivAccum}}), entropy (\code{\link{EntAccum}}) or the Mixing index (\code{\link{Mixing}}).}

\item{...}{Further plotting arguments.}

\item{q}{The order of Diversity. By default, the first value found in the "Accumulation" object is used.}

\item{type}{Plotting parameter. Default is "l".}

\item{main}{Main title of the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{Limits of the Y-axis, as a vector of two numeric values.}

\item{lineH0}{if \code{TRUE}, the values of the null hypothesis are plotted.}

\item{LineWidth}{Width of the Diversity Accumulation Curve line.}

\item{ShadeColor}{The color of the shaded confidence envelope.}

\item{BorderColor}{The color of the borders of the confidence envelope.}
}
\description{
Plot Diversity Accumulation
}
\examples{
# Generate a random community
spCommunity <- rSpCommunity(1, size=50, S=3)
# Calculate mixing indices of order 0 and 1
accum <- Mixing(spCommunity, q.seq=c(0,1))
plot(accum, q=0)
}
