% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shimatani.R
\name{Simpson_r}
\alias{Simpson_r}
\title{Title}
\usage{
Simpson_r(
  X,
  r = NULL,
  correction = "isotropic",
  Biased = TRUE,
  CheckArguments = TRUE
)
}
\arguments{
\item{X}{}

\item{r}{}

\item{correction}{}

\item{Biased}{}

\item{CheckArguments}{}
}
\value{

}
\description{
Title
}
