% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SpatDiv}
\alias{SpatDiv}
\title{SpatDiv}
\description{
Spatially Explicit Measures of Diversity
}
\details{
This package extends the \strong{entropart} package \insertCite{Marcon2014c}{SpatDiv}.
It provides spatially explicit measures of diversity such as the mixing index.
}
\references{
\insertAllCited{}
}
