% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesDistribution.R
\name{SpeciesDistributions}
\alias{SpeciesDistributions}
\alias{as.SpeciesDistribution.wmppp}
\alias{as.ProbaVector.wmppp}
\alias{as.AbdVector.wmppp}
\title{Species Distribution}
\usage{
\method{as.SpeciesDistribution}{wmppp}(x)

\method{as.ProbaVector}{wmppp}(x, Correction = "None", Unveiling = "None",
  RCorrection = "Chao1", JackOver = FALSE, CEstimator = "ZhangHuang",
  CheckArguments = TRUE)

\method{as.AbdVector}{wmppp}(x, Round = TRUE)
}
\arguments{
\item{x}{An object of class "wmppp" (\code{\link{wmppp.object}}), with \code{PointType} values as species names.}

\item{Correction}{A string containing one of the possible corrections to estimate a probability distribution: \code{"None"} (no correction, the default value), \code{"Chao2013"}, \code{"Chao2015"} or \code{"ChaoShen"}.}

\item{Unveiling}{A string containing one of the possible unveiling methods to estimate the probabilities of the unobserved species: \code{"None"} (default, no species is added), \code{"unif"} (uniform: all unobserved species have the same probability) or \code{"geom"} (geometric: the unobserved species distribution is geometric).}

\item{RCorrection}{A string containing a correction recognized by \code{\link{Richness}} to evaluate the total number of species. \code{"Chao1"} is  the default value.}

\item{JackOver}{If \code{TRUE}, retain the jackknife order immediately superior to the optimal one, usually resulting in the overestimation of the number of species. Default is \code{FALSE}. Ignored if \code{RCorrection} is not \code{"Jackknife"}.}

\item{CEstimator}{A string containing an estimator recognized by \code{\link{Coverage}} to evaluate the sample coverage. \code{"ZhangHuang"} is the default value.}

\item{CheckArguments}{Logical; if \code{TRUE}, the function arguments are verified. Should be set to \code{FALSE} to save time when the arguments have been checked elsewhere.}

\item{Round}{Ignored since abundances are obtained by counting points.}
}
\value{
A vector of species abundances (\code{\link{AbdVector}}) or probabilities (\code{\link{ProbaVector}}).
}
\description{
A Species Distribution is a (preferably named) vector containing species abundances or probabilities.
\code{SpeciesDistribution} objects include \code{AbdVector} and \code{ProbaVector} objects.
}
\details{
\code{as.AbdVector} counts the number of individuals (points) per species (in \code{marks$PointType}).

\code{as.ProbaVector} normalizes the vector so that it sums to 1. If \code{Correction} is not \code{"None"}, the observed abundance distribution is used to estimate the actual species distribution. The list of species will be changed: zero-abundance species will be cleared, and some unobserved species will be added. First, observed species probabilities are estimated folllowing Chao and Shen (2003), \emph{i.e.} input probabilities are multiplied by the sample coverage, or according to more sophisticated models: Chao \emph{et al.} (2013, single-parameter model), or Chao \emph{et al.} (2015, two-parameter model). The total probability of observed species equals the sample coverage. Then, the distribution of unobserved species can be unveiled: their number is estimated according to \code{RCorrection} (if the Jackknife estimator is chosen, the \code{JackOver} argument allows using the order immediately over the optimal one). The coverage deficit (1 minus the sample coverage) is shared by the unobserved species equally (\code{Unveiling = "unif"}, Chao \emph{et al.}, 2013) or according to a geometric distribution (\code{Unveiling = "geom"}, Chao \emph{et al.}, 2015).

\code{SpeciesDistribution} objects can be plotted. The \code{plot} method returns the estimated parameters of the fitted distribution. The broken stick has no parameter, so the maximum abundance is returned.
}
